Change beer datatype to decimal.

+++ db/migrate/20150401091637_create_beers.rb
@@ -3,7 +3,7 @@ class CreateBeers < ActiveRecord::Migration
     create_table :beers do |t|
       t.string :name, null: false, unique: true
       t.string :category, null: false
-      t.integer :alc_percent, null: false
+      t.decimal :alc_percent, null: false
       t.string :description
       t.references :user, null: false
